Split EventsWorkerStore into separate file
